Rename classes for better discoverability / grouping
