+++ osu.Game/Screens/SelectV2/PanelBeatmapSet.cs
@@ -19,11 +19,11 @@ using osuTK;
 
 namespace osu.Game.Screens.SelectV2
 {
-    public partial class BeatmapSetPanel : PoolableDrawable, ICarouselPanel
+    public partial class PanelBeatmapSet : PoolableDrawable, ICarouselPanel
     {
         public const float HEIGHT = CarouselItem.DEFAULT_HEIGHT * 1.6f;
 
-        // todo: this should be replaced with information from CarouselItem about how deep is BeatmapPanel in the carousel
+        // todo: this should be replaced with information from CarouselItem about how deep is PanelBeatmap in the carousel
         // (i.e. whether it's under a beatmap set that's under a group, or just under a top-level beatmap set).
         private const float set_x_offset = 20f; // constant X offset for beatmap set/standalone panels specifically.
